move constants from new class to existing SplunkConfiguration
