--- bootstrap/src/main/java/com/splunk/opentelemetry/javaagent/ConfigurationConstants.java
@@ -1,22 +1,0 @@
-/*
- * Copyright Splunk Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.splunk.opentelemetry.javaagent;
-
-public class ConfigurationConstants {
-  public static final String SPLUNK_ACCESS_TOKEN = "splunk.access.token";
-  public static final String PROFILER_ENABLED_PROPERTY = "splunk.profiler.enabled";
-}
