+++ custom/src/main/java/com/splunk/opentelemetry/micrometer/SplunkMetricsConfig.java
@@ -16,8 +16,8 @@
 
 package com.splunk.opentelemetry.micrometer;
 
+import static com.splunk.opentelemetry.SplunkConfiguration.PROFILER_ENABLED_PROPERTY;
 import static com.splunk.opentelemetry.SplunkConfiguration.SPLUNK_ACCESS_TOKEN;
-import static com.splunk.opentelemetry.javaagent.ConfigurationConstants.PROFILER_ENABLED_PROPERTY;
 import static io.micrometer.core.instrument.config.MeterRegistryConfigValidator.checkAll;
 
 import io.micrometer.core.instrument.config.validate.Validated;
